✨ feat(provider): 集成 Command Code 供应商并支持 Vendor Options
- 实现 Command Code 协议适配器，支持 NDJSON 流式解码与 CLI 信封编码
- 引入 Vendor Options 机制，允许供应商定义非密钥行为开关（如 ZDR 零数据保留）
- 优化 Provider Catalog 流程，支持内置供应商并提供结构化的过期选择错误码
- 修复 OpenAI 兼容协议在处理独立推理项时的内容缺失问题，防止上游 400 错误
- 升级数据库 Schema 以支持存储 vendor_options 字段
- 完善 WebUI 供应商编辑器，支持动态渲染并配置供应商选项开关

+++ backend/crates/stravia-core/src/proxy/ingress/open_responses/websocket.rs
@@ -605,7 +605,7 @@ async fn serve(
                                 "close",
                                 frame.map_or(Value::Null, |frame| {
                                     serde_json::json!({
-                                        "code": u16::from(frame.code),
+                                        "code": frame.code,
                                         "reason": frame.reason.as_str(),
                                     })
                                 }),
